Former-commit-id: 6eda240b3b6858d00ebd1406140cf71b7aa2d29f

+++ Programs/UIViewer/Sources/UIViewerApp.cpp
@@ -112,7 +112,7 @@ void UIViewerApp::OnActionTriggered(DAVA::Action action)
         keyboardShowed = !keyboardShowed;
         if (keyboardShowed)
         {
-            w->visibleFrameChanged.Emit(w, Rect(0, 0, w->GetSize().dx, virtualKeyboardHeight));
+            w->visibleFrameChanged.Emit(w, Rect(0, 0, w->GetSize().dx, static_cast<float32>(virtualKeyboardHeight)));
         }
         else
         {
